bpf6: todo: add MTU in arp entry in order to check complement FIB lookup
If the MTU on the interface to redirect is lower than min IPv6
possible, 1280, the fib lookup returns BPF_FIB_LKUP_RET_NOT_FWDED
without updating the mtu filed. To check if the it's a MTU issue
use the ND map to get the if for the destination IP and then check
the MTU for that interface.