bpf6: introduce the bpf program context to avoid bpf_linker error
In order to avoid the bpf_linker errors due to stack memory depletion
every time a new big object is created on the stack or a log call is added
a program context was introduced. This is a per CPU scratchpad memory that
is used to build keys and values for accessing maps or any other operation
that can deplete the 512B stack.

This patch also removes the test that prevented from updating a conntrack
entry due to an already existing cache that keeps the last time it was updated.